Add Greenpeace to earth donation and volunteer
Greenpeace uses peaceful protest and creative communication to expose global environmental problems and to promote solutions that are essential to a green and peaceful future. Trump presents a unique and unprecedented challenge to our environment and democracy. Greenpeace has long and successful history of standing up to bullies and fighting for what’s right, and we are scaling up to build a huge and powerful movement to stand up to Trump.